Renamed optional to optionals to mirror ACE3, Fixed make.cfg optionals path
