+++ optionals/diagnostic_enable_perf_loop/$PBOPREFIX$.txt
