+++ optionals/diagnostic_enable_perf_loop/config.cpp
