+++ optionals/diagnostic_enable_perf_loop/script_component.hpp
