+++ optionals/enable_auto_xeh/$PBOPREFIX$.TXT
