+++ optionals/enable_auto_xeh/script_component.hpp
@@ -1,0 +1,7 @@
+#define COMPONENT enable_auto_xeh
+#include "\x\cba\addons\main\script_mod.hpp"
+
+#undef REQUIRED_VERSION
+#define REQUIRED_VERSION 1.00
+
+#include "\x\cba\addons\main\script_macros.hpp"
