+++ tools/make.cfg
@@ -44,9 +44,9 @@ prefix = x\cba\addons
 # Default: <work_drive>\<prefix>\addons
 module_root = P:\x\cba\addons
 
-# Set the location where the optional addon source folders (i.e. P:\z\ace\optionals)
+# Set the location where the optional addon source folders (i.e. P:\x\cba\optionals)
 # Default: <work_drive>\<prefix>\optionals
-optionals_root = P:\z\ace\optionals
+optionals_root = P:\x\cba\optionals
 
 
 # If the command-line variable test, the addons built will be copied to the following folder.
